change app layout
